refactor: Moved video recording into it's own component
